Pull floating preferences into its own frame
